backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
boto3==1.28.40
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
pytest==7.4.0
